bump syft version (add package.json, rename bundler to ruby)
