+++ test/integration/corner_cases_test.go
@@ -1,5 +1,3 @@
-//+build integration
-
 package integration
 
 import (
